feat: 正式版发布

+++ fluent-sql-core/src/main/java/group/flyfish/fluent/chain/SQLImpl.java
@@ -369,7 +369,7 @@ final class SQLImpl extends ConcatSegment<SQLImpl> implements SQLOperations, Pre
         }
 
         @Override
-        public Mono<Integer> execute() {
+        public Mono<Long> execute() {
             return SHARED_REACTIVE_OPERATIONS.execute(entity);
         }
     }
